Fix stack handling bug in asm_amd64.S causing test failure
Calling sum8 in TestStackArguments fails due to a bug in asm_amd64.S

The function sum8 passes 8 arguments, where two of them are pushed on the
stack before the function in the shared library is called.
After the call instruction in asm_amd64.S returns, the stack must be restored
to the previous state before the CPU registers are restored. Since the stack
is not restored, invalid values are restored to the CPU registers, leading
the incorrect results.

Fix by restoring the stack pointer when arguments are passed on the stack.